feat: add logout modal to mobile and tablet menu with shared state from header
