+++ src/components/MobileMenu/MobileMenu.jsx
@@ -8,9 +8,8 @@ import AppLink from "../AppLink/AppLink";
 import LogoutIcon from '../../assets/icons/exit.svg?react';
 import divider from '../../assets/icons/divider.svg';
 
-const MobileMenu = ({ closeMenu }) => {
+const MobileMenu = ({ closeMenu, isAuthenticated, openModal }) => {
     const user = useSelector(state => state.user.user);
-    const isAuthenticated = Boolean(user?.name);
     const handleLinkClick = () => closeMenu();
     const navLinkClass = ({ isActive }) =>
         (isActive ? `${s.navLink} ${s.active}` : s.navLink);
@@ -52,7 +51,7 @@ const MobileMenu = ({ closeMenu }) => {
                         Creators
                     </NavLink>
                     {isAuthenticated ? (
-                        <NavLink to='/profile' className={navLinkClass} onClick={handleLinkClick}>
+                        <NavLink to={`/authors/${user.id}`} className={navLinkClass} onClick={handleLinkClick}>
                             My Profile
                         </NavLink>
                     ) : (
@@ -77,7 +76,7 @@ const MobileMenu = ({ closeMenu }) => {
                                     </AppLink>
                                 </div>
                                 <img src={divider} alt="divider" className={s.divider} />
-                                <button className={s.logoutBtn} onClick={handleLinkClick}>
+                                <button className={s.logoutBtn} onClick={openModal}>
                                     <LogoutIcon className={s.logoutIcon} width={24} height={28} />
                                 </button>
                             </div>
@@ -94,7 +93,7 @@ const MobileMenu = ({ closeMenu }) => {
                                     <AppLink variant="fill" size="lg" to="/register" onClick={handleLinkClick}>
                                         Join now
                                     </AppLink>
-                                    </div>
+                                </div>
                             )}
                         </div>
                     </div>
